pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Никита Кузнецов",email = "eudaimonia22@yandex.ru"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fabric (>=3.2.2,<4.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-jose[cryptograthy] (>=3.3.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "twisted (>=24.11.0,<25.0.0)",
    "pytest-twisted (>=1.14.3,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "celery (>=5.4.0,<6.0.0)",
    "gevent (>=24.11.1,<25.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "flower (>=2.0.1,<3.0.0)",
    "aio-pika (>=9.5.4,<10.0.0)",
    "fastapi-filter[all] (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
